+++ tap_codat/__init__.py
@@ -6,6 +6,7 @@ from singer import utils, metadata
 from singer.catalog import Catalog, CatalogEntry, Schema
 from . import streams as streams_
 from .context import Context
+from .state import sanitize_bookmarks
 
 REQUIRED_CONFIG_KEYS = ["start_date", "api_key"]
 LOGGER = singer.get_logger()
@@ -45,9 +46,11 @@ def check_credentials_are_authorized(ctx):
 def add_stream_to_catalog(catalog, ctx, stream):
     schema_dict = load_schema(ctx, stream.tap_stream_id)
     schema = Schema.from_dict(schema_dict)
-    mdata = metadata.get_standard_metadata(schema_dict,
-                                           key_properties=stream.pk_fields)
+    mdata = metadata.get_standard_metadata(schema_dict, key_properties=stream.pk_fields, replication_method=stream.replication_method)
     mdata = metadata.to_map(mdata)
+
+    if getattr(stream, 'parent_stream', None):
+        mdata = metadata.write(mdata, (), 'parent-tap-stream-id', stream.parent_stream.tap_stream_id)
 
     for field_name in schema_dict['properties'].keys():
         mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
@@ -86,6 +89,7 @@ def sync(ctx):
         load_and_write_schema(ctx, stream)
         stream.sync(ctx)
     ctx.state["currently_syncing"] = None
+    sanitize_bookmarks(ctx.state)
     ctx.write_state()